+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/B_rechtsschutzversicherung.ts
@@ -1,6 +1,6 @@
 import type { ProzesskostenhilfePDF } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import type { ProzesskostenhilfeFormularContext } from "~/domains/prozesskostenhilfe/formular";
-import type { PkhPdfFillFunction } from ".";
+import type { PkhPdfFillFunction } from "..";
 
 export const fillRechtsschutzversicherung: PkhPdfFillFunction = ({
   userData,
