+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/C_unterhaltspflichtige_person.ts
@@ -1,5 +1,5 @@
 import { familyRelationshipMap } from "~/domains/shared/pdf/unterhaltHelpers";
-import type { PkhPdfFillFunction } from ".";
+import type { PkhPdfFillFunction } from "..";
 
 export const fillUnterhaltsanspruch: PkhPdfFillFunction = ({
   userData,
