+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/F_abzuege.ts
@@ -1,9 +1,9 @@
 import type { AttachmentEntries } from "~/services/pdf/attachment";
 import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/services/pdf/attachment";
 import { pdfFillReducer } from "~/services/pdf/fillOutFunction";
-import type { PkhPdfFillFunction } from ".";
+import type { PkhPdfFillFunction } from "..";
+import { getTotalMonthlyFinancialEntries } from "../util";
 import { zahlungsfrequenzMapping } from "./E_bruttoEinnahmen/bruttoEinnahmen_eigenes";
-import { getTotalMonthlyFinancialEntries } from "./util";
 
 export const fillAbzuege: PkhPdfFillFunction = ({ userData, pdfValues }) => {
   const { pdfValues: filledValues, attachment } = pdfFillReducer({
