+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/H_wohnkosten.ts
@@ -1,4 +1,4 @@
-import type { PkhPdfFillFunction } from ".";
+import type { PkhPdfFillFunction } from "..";
 
 export const fillWohnkosten: PkhPdfFillFunction = ({ userData, pdfValues }) => {
   pdfValues.gesamtgroessedesWohnraumsdenSiealleinodergemeinsammitanderenPersonenbewohnen.value =
